requirements.txt
discord.py
python-dotenv
pdf2image